+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next-release
 
+* frontend/backend: Allow admin to see and edit email.
 * frontend: Hide EditDialog when user is not admin (bug fix).
 
 ## v3.2.0 (released on 2020-10-09)